% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_models.R
\name{cm_logistic}
\alias{cm_logistic}
\title{Logistic Regression Model Wrapper}
\usage{
cm_logistic()
}
\value{
A list with \code{fit}, \code{predict_proba}, and
\code{insample_proba} functions.
}
\description{
Creates a model wrapper around \code{glm(family = binomial)} for binary
classification / propensity score estimation.
}
