% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_models.R
\name{cm_ols}
\alias{cm_ols}
\title{OLS Model Wrapper}
\usage{
cm_ols()
}
\value{
A list with \code{fit} and \code{predict} functions.
}
\description{
Creates a model wrapper around \code{lm()} that conforms to the CausalModel
model protocol: a list with \code{fit(X, y)} and \code{predict(model, X)}
functions.
}
