% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{draw}
\alias{draw}
\title{Draw Treatment Assignment}
\usage{
draw(design, n)
}
\arguments{
\item{design}{A design object.}

\item{n}{Number of units.}
}
\value{
Numeric vector of treatment assignments (0/1).
}
\description{
Draw Treatment Assignment
}
