% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interference.R
\name{encode_G_mat}
\alias{encode_G_mat}
\title{Encode neighborhood structure G to integer}
\usage{
encode_G_mat(G, group_struct)
}
\description{
Encode neighborhood structure G to integer
}
\keyword{internal}
