% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{est_via_ancova}
\alias{est_via_ancova}
\alias{est_via_ancova.experimental}
\title{ANCOVA Estimator}
\usage{
est_via_ancova(obj, ...)

\method{est_via_ancova}{experimental}(obj, ...)
}
\arguments{
\item{obj}{An \code{experimental} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{cm_result} object.
}
\description{
Fisher's ANCOVA with treatment-covariate interactions and HC0 robust
standard errors. Regresses Y ~ 1 + Z + X + Z*X.
}
