% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{est_via_dm}
\alias{est_via_dm}
\alias{est_via_dm.experimental}
\title{Difference-in-Means Estimator}
\usage{
est_via_dm(obj, ...)

\method{est_via_dm}{experimental}(obj, ...)
}
\arguments{
\item{obj}{An \code{experimental} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{cm_result} object.
}
\description{
Simple difference in sample means between treated and control groups.
}
