% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observational.R
\name{est_via_ols}
\alias{est_via_ols}
\alias{est_via_ols.observational}
\title{OLS Estimator}
\usage{
est_via_ols(obj, ...)

\method{est_via_ols}{observational}(obj, ...)
}
\arguments{
\item{obj}{An \code{observational} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{cm_result} object.
}
\description{
Estimates ATE via linear regression of Y on (Z, X) with HC0
heteroskedasticity-robust standard errors.
}
