% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{est_via_strata}
\alias{est_via_strata}
\alias{est_via_strata.experimental}
\title{Stratified Estimator}
\usage{
est_via_strata(obj, ...)

\method{est_via_strata}{experimental}(obj, strata, ...)
}
\arguments{
\item{obj}{An \code{experimental} object.}

\item{...}{Additional arguments passed to methods.}

\item{strata}{Integer or factor vector of stratum labels (same length as Y).}
}
\value{
A \code{cm_result} object.
}
\description{
Estimates ATE as a weighted average of within-stratum difference-in-means.
}
