% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R, R/observational.R
\name{estimate.experimental}
\alias{estimate.experimental}
\alias{estimate}
\alias{estimate.observational}
\title{Default Estimation}
\usage{
\method{estimate}{experimental}(obj, ...)

estimate(obj, ...)

\method{estimate}{observational}(obj, ...)
}
\arguments{
\item{obj}{An \code{observational} or \code{experimental} object.}

\item{...}{Additional arguments passed to the selected estimator.}
}
\value{
A \code{cm_result} object.
}
\description{
Uses AIPW for binary treatment and DML for continuous treatment.
}
