% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{generate_clustered_data}
\alias{generate_clustered_data}
\title{Generate Clustered Data with Varying Cluster Sizes}
\usage{
generate_clustered_data(
  clusters_list = c(5000, 5000, 2000),
  group_struct_list = list(c(2, 3, 4), c(3, 4, 5), c(4, 5, 6)),
  tau = 1,
  gamma = c(5, 0, 1)
)
}
\arguments{
\item{clusters_list}{Integer vector. Number of clusters for each structure.}

\item{group_struct_list}{List of integer vectors. Group structures.}

\item{tau}{Numeric. Direct treatment effect. Default 1.}

\item{gamma}{Numeric vector. Spillover parameters.}
}
\value{
A list with concatenated \code{Y}, \code{Z}, \code{X},
\code{cluster_labels}, \code{group_labels}, \code{ingroup_labels},
\code{G}, \code{Xc}.
}
\description{
Combines multiple fixed-cluster datasets with different group structures.
}
