% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{get_balance}
\alias{get_balance}
\title{Compute Rerandomization Balance Criterion}
\usage{
get_balance(Z, X)
}
\arguments{
\item{Z}{Numeric vector. Treatment indicator.}

\item{X}{Numeric matrix. Covariates.}
}
\value{
Numeric scalar. Balance criterion value.
}
\description{
Mahalanobis-type balance criterion from Morgan & Rubin (2015).
\deqn{m = \frac{n_t n_c}{n} (\bar{X}_t - \bar{X}_c)' \Sigma^{-1}
  (\bar{X}_t - \bar{X}_c)}
}
