% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R
\name{get_params_via_obs}
\alias{get_params_via_obs}
\title{Infer Design Parameters from Observed Treatment}
\usage{
get_params_via_obs(design, Z)
}
\arguments{
\item{design}{A design object.}

\item{Z}{Observed treatment vector.}
}
\value{
Updated design object.
}
\description{
Infer Design Parameters from Observed Treatment
}
