% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{test_via_fisher}
\alias{test_via_fisher}
\alias{test_via_fisher.experimental}
\title{Fisher Randomization Test}
\usage{
test_via_fisher(obj, ...)

\method{test_via_fisher}{experimental}(obj, n_draws = 1000, ...)
}
\arguments{
\item{obj}{An \code{experimental} object.}

\item{...}{Additional arguments passed to methods.}

\item{n_draws}{Number of random draws. Default 1000.}
}
\value{
Numeric p-value.
}
\description{
Computes a p-value by comparing the observed test statistic (DM) to
the distribution under random treatment reassignment.
}
